check deletion_protection in model builder step

+++ pkg/deploy/elbv2/load_balancer_synthesizer.go
@@ -63,7 +63,7 @@ func (s *loadBalancerSynthesizer) Synthesize(ctx context.Context) error {
 	for _, sdkLB := range unmatchedSDKLBs {
 		if err := s.lbManager.Delete(ctx, sdkLB); err != nil {
 			errMessage := err.Error()
-			if strings.Contains(errMessage,"OperationNotPermitted") {
+			if strings.Contains(errMessage,"OperationNotPermitted") && strings.Contains(errMessage, "deletion protection") {
 				s.disableDeletionProtection(sdkLB.LoadBalancer)
 				if err = s.lbManager.Delete(ctx, sdkLB); err != nil {
 					return err
